[Alerting] Add shared getRulesClientMockParams test helper
Introduce a `getRulesClientMockParams` helper in the alerting server
`test_utils` module that centralises construction of the `RulesClient`
`ConstructorOptions` used across unit tests. Adding a new rule client
param now only requires updating this single helper instead of every
unit test file.

Migrate the first rules_client/tests suites (mute_all, unmute_all,
unsnooze_instance) to use it.

Verification (scoped to changed files):
- node scripts/eslint: 0 errors
- node scripts/type_check --project .../alerting/tsconfig.json: 0 errors
- node scripts/jest (migrated suites): all pass
